Refactor: Complete requisitions module isolation into blueprint
This commit finalizes the refactoring of the requisitions module.
Building upon the previous partial refactor, this work addresses
your feedback to ensure a clean separation of concerns.

Key changes:
- Blueprint Definition: Adjusted `app/requisiciones/__init__.py`
  to set `url_prefix="/requisiciones"` directly on the blueprint.
  Simplified blueprint registration in `app/__init__.py`.
- Code Cleanup: Ensured `app/__init__.py` and `app/routes.py` are
  free of requisition-specific logic, with `app/__init__.py`
  focusing on the app factory and blueprint registrations.
- Global Import Updates: Corrected imports for constants and forms
  across the project, notably in `tests/test_flujo_requisiciones.py`
  and `app/models.py`. Verified no legacy imports remain.
- Global `url_for` Update: Updated `url_for` calls in templates
  (e.g., `templates/base.html`, `templates/inicio.html`) to point
  to the new `requisiciones` blueprint endpoints.
- Test Verification: Ensured client calls in
  `tests/test_flujo_requisiciones.py` use the correct
  `/requisiciones/...` URL paths.

The requisitions module is now self-contained within `app/requisiciones`,
and all references, routes, and templates have been updated accordingly.
The main application structure is cleaner as a result.
