+++ app/requisiciones/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+requisiciones_bp = Blueprint("requisiciones", __name__, url_prefix="/requisiciones")
+
+from . import routes
