+++ templates/base.html
@@ -156,13 +156,13 @@
                                 <i class="fas fa-home"></i> Inicio
                             </a>
                             {% if current_user.is_authenticated %}
-                                <a class="dropdown-item" href="{{ url_for('main.listar_requisiciones') }}">
+                                <a class="dropdown-item" href="{{ url_for('requisiciones.listar_requisiciones') }}">
                                     <i class="fas fa-tasks"></i> Ver Requisiciones Activas
                                 </a>
-                                <a class="dropdown-item" href="{{ url_for('main.historial_requisiciones') }}">
+                                <a class="dropdown-item" href="{{ url_for('requisiciones.historial_requisiciones') }}">
                                     <i class="fas fa-history"></i> Ver Historial
                                 </a>
-                                <a class="dropdown-item" href="{{ url_for('main.crear_requisicion') }}">
+                                <a class="dropdown-item" href="{{ url_for('requisiciones.crear_requisicion') }}">
                                     <i class="fas fa-plus-circle"></i> Crear Requisición
                                 </a>
                                 {% if current_user.rol_asignado and current_user.rol_asignado.nombre == 'Admin' %}
