+++ templates/inicio.html
@@ -9,13 +9,13 @@
     <p class="lead">Selecciona una opción para continuar:</p>
 
     <div class="mt-5 d-flex justify-content-center flex-wrap">
-        <a href="{{ url_for('main.crear_requisicion') }}" class="btn btn-granja btn-lg m-2">
+        <a href="{{ url_for('requisiciones.crear_requisicion') }}" class="btn btn-granja btn-lg m-2">
             <i class="fas fa-plus-circle"></i> Crear Nueva Requisición
         </a>
-        <a href="{{ url_for('main.listar_requisiciones') }}" class="btn btn-azul btn-lg m-2">
+        <a href="{{ url_for('requisiciones.listar_requisiciones') }}" class="btn btn-azul btn-lg m-2">
             <i class="fas fa-tasks"></i> Ver Requisiciones Activas
         </a>
-        <a href="{{ url_for('main.historial_requisiciones') }}" class="btn btn-azul btn-lg m-2">
+        <a href="{{ url_for('requisiciones.historial_requisiciones') }}" class="btn btn-azul btn-lg m-2">
             <i class="fas fa-history"></i> Ver Historial
         </a>
     </div>
