+++ templates/requisiciones/historial_requisiciones.html
@@ -93,4 +93,3 @@
     <p class="no-requisiciones mt-4">No hay requisiciones en el historial para mostrar según sus permisos y filtros.</p>
 {% endif %}
 {% endblock %}
-
