+++ templates/requisiciones/listar_por_estado.html
@@ -59,4 +59,3 @@
   </nav>
   {% endif %}
 {% endblock %}
-
